requirements.txt
anthropic
tavily-python
httpx
beautifulsoup4
pypdf
python-docx
reportlab
pydantic>=2.0
fastapi
uvicorn
python-multipart
